商品削除機能の修正(1回目)

+++ app/controllers/items_controller.rb
@@ -38,8 +38,10 @@ class ItemsController < ApplicationController
 
   def destroy
     item = Item.find(params[:id])
-    item.destroy
-    redirect_to root_path
+    if current_user.id == item.user_id 
+      item.destroy
+      redirect_to root_path
+    end
   end
 
   private
